Cast AdjustI2SRate result to uint32_t
Fixes #820

+++ src/AudioOutputPDM.cpp
@@ -61,7 +61,7 @@ bool AudioOutputPDM::begin() {
     assert(ESP_OK == i2s_new_channel(&chan_cfg, &_tx_handle, nullptr));
 
     i2s_pdm_tx_config_t pdm_cfg = {
-        .clk_cfg = I2S_PDM_TX_CLK_DEFAULT_CONFIG(AdjustI2SRate(hertz)),
+        .clk_cfg = I2S_PDM_TX_CLK_DEFAULT_CONFIG((uint32_t)AdjustI2SRate(hertz)),
         .slot_cfg = I2S_PDM_TX_SLOT_DEFAULT_CONFIG(I2S_DATA_BIT_WIDTH_16BIT, I2S_SLOT_MODE_STEREO),
         .gpio_cfg = {
             .clk = I2S_GPIO_UNUSED,
